fix: FranchiseStore 수정
Vendor 참고해서 변경
FranchiseStore 패키지 변경
