--- src/main/java/com/fourweekdays/fourweekdays/franchise/FranchiseStatus.java
@@ -1,6 +1,0 @@
-package com.fourweekdays.fourweekdays.franchise;
-
-public enum FranchiseStatus {
-    ACTIVE, INACTIVE
-}
-
