+++ src/main/java/com/fourweekdays/fourweekdays/franchise/model/entity/FranchiseStatus.java
@@ -1,0 +1,6 @@
+package com.fourweekdays.fourweekdays.franchise.model.entity;
+
+public enum FranchiseStatus {
+    ACTIVE, INACTIVE
+}
+
